SSCS-3428 Update the RPC and Region based in hearing
We no longer want to base the RPC and region on the SC number as we
will not have one in future. Instead we already have a mapping between
the hearing venue id and RPC. If we have a hearing use this to populate
the RPC and the region. If we have multiple hearings use the latest.