SSCS-3428 Update the RPC and Region based on hearing
We no longer want to base the RPC and region on the SC number as we
will not have one in future. Initial we base this in the postcode of
the appellant. However they ask for their hearing to be somewhere else
in which case we want to base it on the location of the last hearing.

If the appellant has a hearing lookup their hearing venue id in our
venue to RPC mapping file.