Improve the retry and recovery logic
We only search for cases before creating them if we are retrying or
recovering.
Fix tests.
Add RetryTemplate configuration to use retry and recovery callbacks.
Add backoff policy to other retryable methods.
Remove testing class that was not doing much sense as it's better
covered in the integration test class.